Created very simple page-level memory allocator
This allocator does not actually keep track of the memory regions
it has allocated, because I am instead experimenting with passing
structs to the allocator functions containing the metadata about
where and how much has been allocated.  I did this in the hope
that making the library trivially re-entrant and thread safe
(so long as it is used right and not called twice with the same
entry) would allow this to just work if I end up dealing with
threads later.

In addition, to simplify this initial implementation and in the
hope that I will not need to expand and re-allocate memory regions
often, I do not allocate any memory ahead of time to accommodate
future expansions.  Instead, I mmap a new memory region every
time page_realloc is called, copy over the original region's pages,
and munmap the original pages.  For my limited memory allocation
use case, I'm hoping this will be sufficient.