Comment out adjustment to size to fit page size
I can't recall why I originally commented this out, but I think it
may be because the original size argument is unchanged, therefore
requiring the caller to check whether entry->size is different
from the size passed.  If I want to change this back to expand
the allocation to fit on page boundaries, then I may want to have
it pass a size_t pointer or return the size allocated.

I don't remember,  but there may have been a specific problem that
led to this being commented out.  Unfortunately I don't have any
notes in regards to this.