Changed library to be aware of wrapper around mmap
Since I am using this with a project that puts a wrapper around
mmap, I needed to provide a way for this library to avoid calling
the wrapped version.  This was required because I added spinlocks
preventing multiple threads from entering wrapped functions, so
entering the wrapper a second time would lead to an infinite wait
at the spinlock.